refactor: Renamed generator classes
Renamed generator classes

Issue: #32
