+++ MinecraftManhunt/src/main/java/com/yoonicode/minecraftmanhuntplus/item_generation/ItemCreator.java
@@ -3,7 +3,5 @@ package com.yoonicode.minecraftmanhuntplus.item_generation;
 import org.bukkit.inventory.ItemStack;
 
 public interface ItemCreator {
-
     ItemStack getItemStack();
-
 }
